% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{na.remove}
\alias{na.remove}
\title{Transform "" levels as NA in factor}
\usage{
na.remove(df)
}
\arguments{
\item{data}{is a dataset}
}
\value{
A updated vector with NA value for NULL values such as "N/A"
}
\description{
Calculate percentages for categorical variables with frequencies
}
\keyword{internal}
